Modified GC so deleted data can be dropped #559
The Oracle is now the authority on GC.  The Oracle looks in Zookeeper to
determine the oldest active timestamp for all transactors.   After finding the
oldest active timestamp if persist this in zookeeper for use by garbage
collection iterators.  The Oracle also returns the GC timestamp it persisted to
Zookeeper to any clients that request a timestamp.  Clients use this GC time
from the oracle when committing or closing a snapshot to determine if they
possibly read stale data.

The GC iterator now drops deleted data when its safe to do so.  The reason the
changes were made to the Oracle was so that the GC iterator could stop placing
truncation markers when it dropped data.  Safely getting rid of truncation
markers was necessary in order to drop deleted data.