Fix #871 Use regular jar for ITs module
Move tests in the integration test module from src/test/java into
src/main/java in order to create a regular jar instead of a test-jar in
Maven. This fixes many issues, such as not having javadoc and source
jars for the integration test module, as well as to simplify any
dependency resolution issues with other modules or projects depending on
the integration test jar.