Updated Fluo to build and run with Java 11
Fluo was not building and running with Java 11.  Twill was causing
problems with Java 11.  Since twill was part of the built in support for
running Fluo in YARN that was deprecated in Fluo 1.2.0 it was removed
along with the deprected fluo-cluster module, related fluo script
commands, and deprecated fluo properties file. Also some deperecated
methods in the Fluo configuration API had to be dropped.  Because of the
changes to the API and scripts, Fluo's version was bumped to
2.0.0-SNAPSHOT.

Fluo initialization was assuming the System classloader was a URL
classloader.  This assumption is no longer true in Java 11.  This code
was updated to use the java system property java.class.path.

Findbugs was not working with Java 11, so this was replaced with
spotbugs in the poms.  Some changes were made to fix some issue found by
spotbugs.