Parallelize TransactionImpl.readUnread()
When a transaction only writes to a row+col and then has a collision
Fluo will read the row+col after the collision to look for orphaned
locks. This commit parallelizes this behavior by reading all row+cols
at once.

This commit accomplishes this by using a ParallelSnapshotScanner
instead of a SnapshotScanner.

Fixes #948