Centralize command logic to be run from one class
Using the Command Pattern as inspiration, all the logic for running a
command has been moved to the FluoProgram class. The purpose of this
was to prevent calling System.exit() from within the command classes. 
This will allow the commands to be more easily reused outside of a CLI.
This also makes it easier to print standard error messages for expected
exceptions by throwing a FluoCommandException, and to print a stack 
trace for all other exceptions.

Fixes #983