Use LeaderLatch to determine if oracle exists (#1088)
* Use LeaderLatch to determine if oracle exists

The old implementation of oracleExists checks to see if there are
more than 1 ZNodes under the ZookeeperPath.ORACLE_SERVER path. This
relies on the LeaderLatch implementation and uses a path given to
a Curator recipe, which is advised against by the Curator docs.
See https://cwiki.apache.org/confluence/display/CURATOR/TN7.

The new implementation utilizes LeaderLatch to count the number
of participants to see if there's more than 0.