Remove race condition from FluoAdminImpl.numWorkers (#1090)
The pattern of checking if a ZNode exists and then getting it's
children contains a race condition. If the ZNode is deleted after
the existence check but before getting the children then a
NoNodeException will be thrown instead of returning 0.

Instead we skip the existence check and just return 0 when a
NoNodeException is thrown. This will avoid the race condition.

Fixes #1089